Optimizer refactoring (#1050)
* Optimizer declaration moved to separate header file

* Added override input and output parser for optimizer pass
