+++ include/ttmlir/Dialect/TTNN/Transforms/Passes.td
@@ -21,34 +21,4 @@ def TTNNDecomposeCompositeLayouts: Pass<"ttnn-decompose-composite-layouts", "::m
   }];
 }
 
-def TTNNOptimizer: Pass<"ttnn-optimizer", "::mlir::ModuleOp"> {
-  let summary = "Determine op configurations for maximum performance.";
-  let description = [{
-    Go through the ops, set sharding specs for each op based on sharding analysis,
-    by updating layout attribute of each op.
-  }];
-  let options = [
-    Option<"overrideInputLayout", "insert-reshard",
-          "llvm::StringMap<InputLayoutOverrideParams>",
-          /*default=*/"llvm::StringMap<InputLayoutOverrideParams>()",
-           "Manually insert reshard for specific op's operand.">,
-    Option<"overrideOutputLayout", "override-output-layout",
-          "llvm::StringMap<OutputLayoutOverrideParams>",
-          /*default=*/"llvm::StringMap<OutputLayoutOverrideParams>()",
-           "Override output tensor layout for specific ops.">,
-    Option<"shardingPassEnabled", "sharding-pass-enabled",
-          "bool",
-          /*default=*/"false",
-           "Enable sharding pass.">,
-    Option<"reshardingEnabled", "resharding-enabled",
-          "bool",
-          /*default=*/"false",
-          "Resharding pass. Temp disabled till we support all types of shard specs.">,
-    Option<"maxLegalLayouts", "max-legal-layouts",
-          "int64_t",
-          /*default=*/"64",
-          "Override maximum number of legal layouts for grid analysis.">
-  ];
-}
-
 #endif
